requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
gitpython
sentence-transformers
ollama
python-dotenv